Refactor CSVFileRepository.
