--- app/src/ereferrals/java/org/eyeseetea/malariacare/domain/boundary/repositories/ICSVFileRepository.java
@@ -1,12 +1,0 @@
-package org.eyeseetea.malariacare.domain.boundary.repositories;
-
-import org.eyeseetea.malariacare.data.IDataSourceCallback;
-
-public interface ICSVFileRepository {
-
-    void getCSVFile(String filename, IDataSourceCallback<byte[]> callback);
-
-    void saveCSVFile(String filename, byte[] fileToSave, IDataSourceCallback<Void> callback);
-
-    void updateCSVDB(IDataSourceCallback<Void> callback);
-}
